Second pass: wrapping all mysql endpoints
I'd have refactored and cleaned this up a bit, but there are no tests. I
suppose I may have broken some of these endpoints. If there are any
endpoints that matter they matter to yeti's and will break the yeti's,
leading to this comment and a brick of shame for me.